# *: optimize SeekPrefixGE to use Next