CockroachDB uses SeekPrefixGE on the Put path to read
the existing value for a key. For a batch of Puts, the
common case is to use the same Iterator with a sequence
of SeekPrefixGE calls (with no intermediate Next calls).
These seek calls use keys that are sorted in ascending
order.

For a CockroachDB benchmark that does
`UPDATE bench.kv SET v = v + 1 WHERE k IN (...)`
for a batch of 100 keys, we see the following benefit

```
name                                old time/op    new time/op    delta
KVAndStorageUsingSQL/versions=2-16    3.27ms ± 2%    2.07ms ± 2%  -36.69%  (p=0.008 n=5+5)
KVAndStorageUsingSQL/versions=4-16    3.28ms ± 1%    2.07ms ± 0%  -36.90%  (p=0.008 n=5+5)
KVAndStorageUsingSQL/versions=8-16    3.65ms ± 1%    2.44ms ± 1%  -33.01%  (p=0.008 n=5+5)
```

Note that the above query also does a batch of 100 scans
to read the value for each key (the put is not
responsible for computing v + 1), which is not being
optimized here (it is already optimized to use Next).

This optimization is also needed to recover
the performance losses caused by using separated
intents as discussed in detail in
cockroachdb/cockroach#59397

Additionally, this PR optimizes away the allocation
of the Iterator.prefix byte slice.

The following are results from new benchmarks.
The arenaskl.Iterator and levelIter benchmarks show
speedup when this optimization is successful and small
slowdowns when it is not (higher skip settings). The
mergingIter shows speedup even with higher skip since
there are many sparse levels where we are already
beyond the next seek key, which benefit from this
optimization.

arenaskl.Iterator benchmark
```
BenchmarkSeekPrefixGE/skip=1/use-next=false-16         	 3059583	       388 ns/op	      22 B/op	       2 allocs/op
BenchmarkSeekPrefixGE/skip=1/use-next=true-16          	 7317049	       157 ns/op	      22 B/op	       2 allocs/op
BenchmarkSeekPrefixGE/skip=2/use-next=false-16         	 3095071	       382 ns/op	      22 B/op	       2 allocs/op
BenchmarkSeekPrefixGE/skip=2/use-next=true-16          	 7035187	       173 ns/op	      22 B/op	       2 allocs/op
BenchmarkSeekPrefixGE/skip=4/use-next=false-16         	 3015534	       398 ns/op	      22 B/op	       2 allocs/op
BenchmarkSeekPrefixGE/skip=4/use-next=true-16          	 5753680	       211 ns/op	      22 B/op	       2 allocs/op
BenchmarkSeekPrefixGE/skip=8/use-next=false-16         	 3046408	       403 ns/op	      22 B/op	       2 allocs/op
BenchmarkSeekPrefixGE/skip=8/use-next=true-16          	 2495090	       488 ns/op	      22 B/op	       2 allocs/op
BenchmarkSeekPrefixGE/skip=16/use-next=false-16        	 2754469	       419 ns/op	      22 B/op	       2 allocs/op
BenchmarkSeekPrefixGE/skip=16/use-next=true-16         	 2462174	       484 ns/op	      22 B/op	       2 allocs/op
```

levelIter benchmark
```
BenchmarkLevelIterSeqSeekPrefixGE/skip=1/use-next=false-16         	 2033031	       593 ns/op
BenchmarkLevelIterSeqSeekPrefixGE/skip=1/use-next=true-16          	12691249	        95.3 ns/op
BenchmarkLevelIterSeqSeekPrefixGE/skip=2/use-next=false-16         	 2035592	       593 ns/op
BenchmarkLevelIterSeqSeekPrefixGE/skip=2/use-next=true-16          	10188423	       117 ns/op
BenchmarkLevelIterSeqSeekPrefixGE/skip=4/use-next=false-16         	 2030196	       588 ns/op
BenchmarkLevelIterSeqSeekPrefixGE/skip=4/use-next=true-16          	 6742580	       173 ns/op
BenchmarkLevelIterSeqSeekPrefixGE/skip=8/use-next=false-16         	 1827902	       712 ns/op
BenchmarkLevelIterSeqSeekPrefixGE/skip=8/use-next=true-16          	 1454386	       761 ns/op
BenchmarkLevelIterSeqSeekPrefixGE/skip=16/use-next=false-16        	 1442826	       833 ns/op
BenchmarkLevelIterSeqSeekPrefixGE/skip=16/use-next=true-16         	 1280570	       936 ns/op
```

mergingIter benchmark
```
BenchmarkMergingIterSeqSeekPrefixGE/skip=1/use-next=false-16         	  700455	      1710 ns/op	       0 B/op	       0 allocs/op
BenchmarkMergingIterSeqSeekPrefixGE/skip=1/use-next=true-16          	 2429803	       505 ns/op	       0 B/op	       0 allocs/op
BenchmarkMergingIterSeqSeekPrefixGE/skip=2/use-next=false-16         	  683803	      1694 ns/op	       0 B/op	       0 allocs/op
BenchmarkMergingIterSeqSeekPrefixGE/skip=2/use-next=true-16          	 2300224	       520 ns/op	       0 B/op	       0 allocs/op
BenchmarkMergingIterSeqSeekPrefixGE/skip=4/use-next=false-16         	  668989	      1751 ns/op	       0 B/op	       0 allocs/op
BenchmarkMergingIterSeqSeekPrefixGE/skip=4/use-next=true-16          	 2111007	       568 ns/op	       0 B/op	       0 allocs/op
BenchmarkMergingIterSeqSeekPrefixGE/skip=8/use-next=false-16         	  658203	      1817 ns/op	       0 B/op	       0 allocs/op
BenchmarkMergingIterSeqSeekPrefixGE/skip=8/use-next=true-16          	  998536	      1191 ns/op	       0 B/op	       0 allocs/op
BenchmarkMergingIterSeqSeekPrefixGE/skip=16/use-next=false-16        	  620912	      1990 ns/op	       0 B/op	       0 allocs/op
BenchmarkMergingIterSeqSeekPrefixGE/skip=16/use-next=true-16         	  856705	      1343 ns/op	       0 B/op	       0 allocs/op
```